Clase 5 - Material

+++ src/app/person/person-list/persons-list.component.ts
@@ -4,6 +4,8 @@ import {PersonService} from "../../services/person.service";
 import {Router} from "@angular/router";
 import {PersonGenericService} from "../../services/person-generic.service";
 import {map, Subscription} from "rxjs";
+import {MatDialog} from "@angular/material/dialog";
+import {ConfirmationComponent} from "../../shared/confirmation/confirmation.component";
 
 @Component({
   selector: 'app-person-list',
@@ -14,9 +16,13 @@ export class PersonsListComponent implements OnInit, OnDestroy {
   persons: Person[] = [];
   personSelected: Person | undefined = undefined;
   sub: Subscription[] = [];
+  loading = true;
+
+  displayedColumns: string[] = ['name', 'lastName', 'age', 'options'];
 
   constructor(private personService: PersonGenericService,
-              private router: Router) {
+              private router: Router,
+              private dialog: MatDialog) {
   }
 
   ngOnInit(): void {
@@ -27,7 +33,9 @@ export class PersonsListComponent implements OnInit, OnDestroy {
   }
 
   findAll() {
+    this.loading = true;
     this.personService.findAll().subscribe(list => {
+      this.loading = false;
       this.persons = list.map(p => new Person(p.id, p.firstName, p.lastName, p.age))
     })
   }
@@ -40,12 +48,17 @@ export class PersonsListComponent implements OnInit, OnDestroy {
   }
 
   delete(p: Person) {
-    this.sub.push(this.personService.delete(p.id).subscribe({
-      next: () => {
-        alert("Persona eliminada con exito")
-        this.findAll()
-      },
-      error: (err) => alert(err),
-    }));
+    const dialogRef = this.dialog.open(ConfirmationComponent)
+    dialogRef.afterClosed().subscribe(res => {
+      if (res) {
+        this.sub.push(this.personService.delete(p.id).subscribe({
+          next: () => {
+            alert("Persona eliminada con exito")
+            this.findAll()
+          },
+          error: (err) => alert(err),
+        }));
+      }
+    })
   }
 }
